# shared-contracts/openapi/Sacco.Api.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "SACCO Platform API",
    "description": "Core banking API for Kenyan Deposit-Taking SACCOs. Pass the tenant slug in the X-Tenant header when not resolving by host. Tokens are issued by the in-process OIDC server at /connect/*.",
    "version": "v1"
  },
  "paths": {
...
    "/api/payments/webhooks/{tenant}/{provider}": {
      "post": {
        "tags": [
          "Provider webhooks"
        ],
        "operationId": "ProviderWebhook",
        "parameters": [
          {
            "name": "tenant",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/admin/tenant/branding": {
      "put": {
        "tags": [
          "Platform admin"
        ],
        "operationId": "UpdateTenantBranding",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBrandingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/admin/audit-log": {
      "get": {
        "tags": [
          "Platform admin"
        ],
        "operationId": "GetAuditLog",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 50
            }
          },
          {
            "name": "entityType",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "entityId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfAuditLogResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/payments/collections": {
      "post": {
        "tags": [
          "Payments"
        ],
        "operationId": "InitiateCollection",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitiateCollectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentTransactionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/payments/disbursements": {
      "post": {
        "tags": [
          "Payments"
        ],
        "operationId": "InitiateWithdrawalPayout",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitiateDisbursementRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentTransactionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/payments/transactions": {
      "get": {
        "tags": [
          "Payments"
        ],
        "operationId": "ListPaymentTransactions",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/PaymentStatus"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/PaymentKind"
            }
          },
          {
            "name": "provider",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfPaymentTransactionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/payments/transactions/{id}": {
      "get": {
        "tags": [
          "Payments"
        ],
        "operationId": "GetPaymentTransaction",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentTransactionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/payments/transactions/{id}/reconcile": {
      "post": {
        "tags": [
          "Payments"
        ],
        "operationId": "ReconcilePaymentTransaction",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentTransactionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/payments/sandbox/transactions/{id}/callback": {
      "post": {
        "tags": [
          "Payments"
        ],
        "operationId": "SimulateSandboxCallback",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "success",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnonymousTypeOfstringAndSandboxCallback"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnonymousTypeOfstringAndSandboxCallback": {
        "required": [
          "message",
          "callback"
        ],
        "type": "object",
        "properties": {
          "message": {
            "type": [
              "null",
              "string"
            ]
          },
          "callback": {
            "$ref": "#/components/schemas/SandboxCallback"
          }
        }
      },
...
      "AuditLogResponse": {
        "required": [
          "id",
          "occurredAt",
          "action",
          "entityType",
          "entityId",
          "actorUserId",
          "details",
          "correlationId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "occurredAt": {
            "type": "string",
            "format": "date-time"
          },
          "action": {
            "type": "string"
          },
          "entityType": {
            "type": "string"
          },
          "entityId": {
            "type": "string"
          },
          "actorUserId": {
            "type": "string",
            "format": "uuid"
          },
          "details": {
            "type": [
              "null",
              "string"
            ]
          },
          "correlationId": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "InitiateCollectionRequest": {
        "required": [
          "provider",
          "phoneNumber",
          "amount",
          "purpose",
          "accountNumber",
          "loanNumber",
          "narrative"
        ],
        "type": "object",
        "properties": {
          "provider": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "amount": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "purpose": {
            "$ref": "#/components/schemas/PaymentPurposeType"
          },
          "accountNumber": {
            "type": [
              "null",
              "string"
            ]
          },
          "loanNumber": {
            "type": [
              "null",
              "string"
            ]
          },
          "narrative": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "InitiateDisbursementRequest": {
        "required": [
          "withdrawalId",
          "provider"
        ],
        "type": "object",
        "properties": {
          "withdrawalId": {
            "type": "string",
            "format": "uuid"
          },
          "provider": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "PagedResultOfAuditLogResponse": {
        "required": [
          "items",
          "page",
          "pageSize",
          "totalCount"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuditLogResponse"
            }
          },
          "page": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "pageSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "totalCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "PagedResultOfPaymentTransactionResponse": {
        "required": [
          "items",
          "page",
          "pageSize",
          "totalCount"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PaymentTransactionResponse"
            }
          },
          "page": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "pageSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "totalCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "PaymentKind": {
        "enum": [
          "Collection",
          "Disbursement"
        ]
      },
      "PaymentPurposeType": {
        "enum": [
          "SavingsDeposit",
          "LoanRepayment",
          "WithdrawalPayout",
          "Unallocated"
        ]
      },
      "PaymentStatus": {
        "enum": [
          "Initiated",
          "PendingCallback",
          "Succeeded",
          "Failed",
          "TimedOut"
        ]
      },
      "PaymentTransactionResponse": {
        "required": [
          "id",
          "kind",
          "provider",
          "status",
          "amount",
          "counterparty",
          "memberId",
          "purposeType",
          "accountNumber",
          "loanNumber",
          "withdrawalId",
          "ourReference",
          "providerRequestId",
          "providerTransactionReference",
          "failureReason",
          "ledgerJournalEntryId",
          "initiatedByUserId",
          "initiatedAt",
          "completedAt",
          "callbackCount",
          "narrative"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "kind": {
            "$ref": "#/components/schemas/PaymentKind"
          },
          "provider": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/PaymentStatus"
          },
          "amount": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "counterparty": {
            "type": "string"
          },
          "memberId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "purposeType": {
            "$ref": "#/components/schemas/PaymentPurposeType"
          },
          "accountNumber": {
            "type": [
              "null",
              "string"
            ]
          },
          "loanNumber": {
            "type": [
              "null",
              "string"
            ]
          },
          "withdrawalId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "ourReference": {
            "type": "string"
          },
          "providerRequestId": {
            "type": [
              "null",
              "string"
            ]
          },
          "providerTransactionReference": {
            "type": [
              "null",
              "string"
            ]
          },
          "failureReason": {
            "type": [
              "null",
              "string"
            ]
          },
          "ledgerJournalEntryId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "initiatedByUserId": {
            "type": "string",
            "format": "uuid"
          },
          "initiatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "completedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "callbackCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "narrative": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "SandboxCallback": {
        "required": [
          "providerRequestId",
          "providerTransactionReference",
          "result",
          "amount",
          "phoneNumber",
          "accountReference",
          "failureReason"
        ],
        "type": "object",
        "properties": {
          "providerRequestId": {
            "type": [
              "null",
              "string"
            ]
          },
          "providerTransactionReference": {
            "type": "string"
          },
          "result": {
            "type": "string"
          },
          "amount": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "phoneNumber": {
            "type": [
              "null",
              "string"
            ]
          },
          "accountReference": {
            "type": [
              "null",
              "string"
            ]
          },
          "failureReason": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "UpdateBrandingRequest": {
        "required": [
          "primaryColor",
          "secondaryColor",
          "accentColor",
          "logoUrl",
          "tagline",
          "supportEmail",
          "supportPhone"
        ],
        "type": "object",
        "properties": {
          "primaryColor": {
            "type": "string"
          },
          "secondaryColor": {
            "type": "string"
          },
          "accentColor": {
            "type": "string"
          },
          "logoUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "tagline": {
            "type": "string"
          },
          "supportEmail": {
            "type": "string"
          },
          "supportPhone": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}